graph: move abstract graph types
Moves the abstract `GraphLike`, `NodeLike` and `EdgeLike` types
from `@vltpkg/graph` to `@vltpkg/types` to avoid circular dependencies.

Refs: https://github.com/vltpkg/vltpkg/issues/748
