+++ src/graph/src/browser.ts
@@ -4,7 +4,6 @@ import { getBooleanFlagsFromNum } from './lockfile/types.ts'
 import { stringifyNode } from './stringify-node.ts'
 import { loadEdges } from './lockfile/load-edges.ts'
 import { loadNodes } from './lockfile/load-nodes.ts'
-export * from './types.ts'
 
 const lockfile = {
   loadEdges,
