+++ src/graph/src/index.ts
@@ -7,7 +7,6 @@ export * from './visualization/json-output.ts'
 export * from './visualization/human-readable-output.ts'
 export * from './visualization/mermaid-output.ts'
 export * from './stringify-node.ts'
-export * from './types.ts'
 export * from './install.ts'
 export * from './uninstall.ts'
 export * from './update.ts'
