+++ src/graph/src/lockfile/load-nodes.ts
@@ -2,7 +2,7 @@ import { splitDepID } from '@vltpkg/dep-id/browser'
 import { getBooleanFlagsFromNum } from './types.ts'
 import type { LockfileData, LockfileNode } from './types.ts'
 import type { DepID } from '@vltpkg/dep-id'
-import type { GraphLike } from '../types.ts'
+import type { GraphLike } from '@vltpkg/types'
 
 export const loadNodes = (
   graph: GraphLike,
