--- src/graph/src/types.ts
@@ -1,87 +1,0 @@
-import type { DepID } from '@vltpkg/dep-id'
-import type {
-  NormalizedManifest,
-  DependencyTypeShort,
-} from '@vltpkg/types'
-import type { Spec, SpecLikeBase } from '@vltpkg/spec'
-
-export type EdgeLike = {
-  name: string
-  from: NodeLike
-  spec: SpecLikeBase
-  to?: NodeLike
-  type: DependencyTypeShort
-  optional?: boolean
-  peer?: boolean
-}
-
-export type GraphLike = {
-  importers: Set<NodeLike>
-  mainImporter: NodeLike
-  projectRoot: string
-  nodes: Map<DepID, NodeLike>
-  nodesByName: Map<string, Set<NodeLike>>
-  edges: Set<EdgeLike>
-  addEdge: (
-    type: DependencyTypeShort,
-    spec: Spec,
-    from: NodeLike,
-    to?: NodeLike,
-  ) => EdgeLike
-  addNode: (
-    id?: DepID,
-    manifest?: NormalizedManifest,
-    spec?: Spec,
-    name?: string,
-    version?: string,
-  ) => NodeLike
-}
-
-export type NodeLike = {
-  id: DepID
-  confused: boolean
-  edgesIn: Set<EdgeLike>
-  edgesOut: Map<string, EdgeLike>
-  location?: string
-  manifest?: NormalizedManifest | null
-  rawManifest?: NormalizedManifest | null
-  name?: string | null
-  version?: string | null
-  integrity?: string | null
-  resolved?: string | null
-  importer: boolean
-  graph: GraphLike
-  mainImporter: boolean
-  projectRoot: string
-  dev: boolean
-  optional: boolean
-  modifier?: string | undefined
-  registry?: string
-  toJSON: () => Pick<
-    NodeLike,
-    | 'id'
-    | 'name'
-    | 'version'
-    | 'location'
-    | 'importer'
-    | 'manifest'
-    | 'projectRoot'
-    | 'integrity'
-    | 'resolved'
-    | 'dev'
-    | 'optional'
-    | 'confused'
-  > & {
-    rawManifest?: NodeLike['manifest']
-  }
-  toString(): string
-  setResolved(): void
-  setConfusedManifest(
-    fixed: NormalizedManifest,
-    confused?: NormalizedManifest,
-  ): void
-  maybeSetConfusedManifest(
-    spec: Spec,
-    confused?: NormalizedManifest,
-  ): void
-}
