+++ src/graph/src/visualization/mermaid-output.ts
@@ -1,7 +1,7 @@
 import type { DepID } from '@vltpkg/dep-id'
 import { Edge } from '../edge.ts'
 import { Node } from '../node.ts'
-import type { EdgeLike, NodeLike } from '../types.ts'
+import type { EdgeLike, NodeLike } from '@vltpkg/types'
 
 export type MermaidOutputGraph = {
   edges: EdgeLike[]
