+++ src/graph/test/graph.ts
@@ -7,7 +7,7 @@ import { inspect } from 'node:util'
 import t from 'tap'
 import { Edge } from '../src/edge.ts'
 import { Graph } from '../src/graph.ts'
-import type { GraphLike } from '../src/types.ts'
+import type { GraphLike } from '@vltpkg/types'
 
 t.cleanSnapshot = s =>
   s.replace(/^(\s+)"projectRoot": .*$/gm, '$1"projectRoot": #')
