+++ src/graph/test/node.ts
@@ -5,7 +5,7 @@ import { inspect } from 'node:util'
 import t from 'tap'
 import { Edge } from '../src/edge.ts'
 import { asNode, isNode, Node } from '../src/node.ts'
-import type { GraphLike } from '../src/types.ts'
+import type { GraphLike } from '@vltpkg/types'
 import { PathScurry } from 'path-scurry'
 
 t.cleanSnapshot = s =>
