+++ src/query/src/pseudo.ts
@@ -59,7 +59,7 @@ import { unpopular } from './pseudo/unpopular.ts'
 import { unstable } from './pseudo/unstable.ts'
 import { workspace } from './pseudo/workspace.ts'
 
-import type { EdgeLike, NodeLike } from '@vltpkg/graph'
+import type { EdgeLike, NodeLike } from '@vltpkg/types'
 import type { ParserFn, ParserState } from './types.ts'
 
 /**
