+++ src/query/test/combinator.ts
@@ -8,7 +8,7 @@ import {
   selectorFixture,
 } from './fixtures/selector.ts'
 import type { TestCase } from './fixtures/types.ts'
-import type { EdgeLike, NodeLike } from '@vltpkg/graph'
+import type { EdgeLike, NodeLike } from '@vltpkg/types'
 import type { GraphSelectionState } from '../src/types.ts'
 
 const testCombinator = selectorFixture(combinator)
