+++ src/query/test/fixtures/graph.ts
@@ -1,9 +1,8 @@
 import { joinDepIDTuple } from '@vltpkg/dep-id'
 import { Spec } from '@vltpkg/spec/browser'
 import { normalizeManifest } from '@vltpkg/types'
-import type { GraphLike, NodeLike } from '@vltpkg/graph'
+import type { GraphLike, NodeLike , Manifest, DependencyTypeShort } from '@vltpkg/types'
 import type { SpecLike, SpecOptions } from '@vltpkg/spec/browser'
-import type { Manifest, DependencyTypeShort } from '@vltpkg/types'
 
 const specOptions = {
   registries: {
